+++ controller/controllerQuestion.py
@@ -18,29 +18,38 @@ class ControllerQuestion:
             return self.__dao.insert(question)
 
     def update(self, id: int, description=None, answer=None, id_category=None, points=None, date=None):
-        question = self.__dao.read(id)
+        if isinstance(id, int):
+            question = self.__dao.read(id)
 
-        if description:
-            question.description = description
-        if answer:
-            question.answer = answer
-        if id_category:
-            category = self.__controller_category.read(id_category)
-            if category:
-                question.category = category
-        if points:
-            question.points = points
-        if date:
-            question.date = date
+            if description:
+                question.description = description
+            if answer:
+                question.answer = answer
+            if id_category:
+                category = self.__controller_category.read(id_category)
+                if category:
+                    question.category = category
+            if points:
+                question.points = points
+            if date:
+                question.date = date
 
-        return self.__dao.update(question)
+            return self.__dao.update(question)
+        else:
+            raise TypeError
 
     def delete(self, id: int):
-        question = self.__dao.read(id)
-        return self.__dao.delete(question)
+        if isinstance(id, int):
+            question = self.__dao.read(id)
+            return self.__dao.delete(question)
+        else:
+            raise TypeError
 
     def read(self, id: int):
-        return self.__dao.read(id)
+        if isinstance(id, int):
+            return self.__dao.read(id)
+        else:
+            raise TypeError
 
     def readRandom(self):
         list = self.__dao.list()
